Fix Dockerfile to avoid persisting ENV variables
We must take care about persisting environment vairables in the
image that are not really needed at this point, because those
variables (former build_type) won't be possible to be used on
images inheriting from this image.

Also, we rename to upper case, the real environment variables
needed by the builder script and which will be provided at
docker run when needed.

We keep lower case `make_procs` to adapt the current image
creation to the environment where it is created.

We also add the usage section to README file.

Change-Id: I2eda86f87aef5857c3355932081a571bc80f1629