Increase max concurrent streams to 32768
nghttp2 does not allow to configure this value through
the API, so we need to patch the source code to a fixed
value. We consider 2^15 good enough to permit server
applications to manage delayed contexts with good
throughput.

For example, in h2agent [0] if we provision a resonse delay
of 2 seconds, with previous maximum of 100 concurrent streams,
we will have only 50 reqs/seconds for a single instance.

With this value of 32768, theoretically we will have 16k reqs/second
with such delay, which is the current benchmarking for non-delay with
h2load in a 8-vcpu vagrant bionic machine.

Also, we upgrade nghttp2 and boost versions:

nghttp2: 1.42.0 -> 1.45.1 (*)
boost: 1.72.0 -> 1.76.0

(*) we must add '--enable-python-bindings=no' to build the new version
    due to a known bug in project configuration.