Downgrade boost version to 1.76.0
Seems that nghttp2 server is crashing after reception of http2
request, when it is linked to boost version over 1.76.0.

This is reported here [0].

Until solved, we will stuck on 1.76.0 version, which
is the newest one without crashing.
This problem only happens to docker-builds (under alpine).
when generating the executable natively (ubuntu) the crash
does not appear.

Also, we improve the patches procedure to avoid failing when no patches
are included for an specific version.

[0] https://github.com/nghttp2/nghttp2/issues/1721