Fix bug in builder make arguments
There is a corner case for builder behavior.
When the project uses cmake, but only provides 1 argument
(which is provided to cmake command), this will be also
passed to make because shift is ignored due to checking
condition for further arguments. So, we need to modify
the checking to shift only when there is at least 1 argument.